Project/requirements.txt
matplotlib==3.9.0
numpy==1.26.4
numpy-stl==3.1.1

